Move code back from graphics_util.h
