+++ engine/graphics/src/graphics_util.h
@@ -53,151 +53,6 @@ namespace dmGraphics
 #endif
         return dmVMath::Vector4(r,g,b,a);
     }
-
-    static inline bool IsColorBufferType(BufferType buffer_type)
-    {
-        return buffer_type == BUFFER_TYPE_COLOR0_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR1_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR2_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR3_BIT;
-    }
-
-    static inline void SetWriteAttributeStreamDesc(WriteAttributeStreamDesc* stream, const float** data, VertexAttribute::VectorType vector_type, uint8_t stream_count, bool is_global_data)
-    {
-        stream->m_Data = data;
-        stream->m_VectorType = vector_type;
-        stream->m_StreamCount = stream_count;
-        stream->m_IsGlobalData = is_global_data;
-    }
-
-    static inline void VertexAttributeInfoMetadataMember(VertexAttributeInfoMetadata& metadata, VertexAttribute::SemanticType semantic_type, CoordinateSpace coordinate_space)
-    {
-        switch(semantic_type)
-        {
-        case VertexAttribute::SEMANTIC_TYPE_POSITION:
-            metadata.m_HasAttributeWorldPosition |= coordinate_space == COORDINATE_SPACE_WORLD;
-            metadata.m_HasAttributeLocalPosition |= coordinate_space == COORDINATE_SPACE_LOCAL;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_TEXCOORD:
-            metadata.m_HasAttributeTexCoord = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_PAGE_INDEX:
-            metadata.m_HasAttributePageIndex = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_NORMAL:
-            metadata.m_HasAttributeNormal = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_TANGENT:
-            metadata.m_HasAttributeTangent = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_COLOR:
-            metadata.m_HasAttributeColor = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_WORLD_MATRIX:
-            metadata.m_HasAttributeWorldMatrix = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_NORMAL_MATRIX:
-            metadata.m_HasAttributeNormalMatrix = true;
-            break;
-        case VertexAttribute::SEMANTIC_TYPE_NONE:
-            metadata.m_HasAttributeNone = true;
-            break;
-        default:
-            break;
-        }
-    }
-
-    static inline VertexAttributeInfoMetadata GetVertexAttributeInfosMetaData(const VertexAttributeInfos& attribute_infos)
-    {
-        VertexAttributeInfoMetadata metadata = {};
-        for (int i = 0; i < attribute_infos.m_NumInfos; ++i)
-        {
-            const VertexAttributeInfo& info = attribute_infos.m_Infos[i];
-            VertexAttributeInfoMetadataMember(metadata, info.m_SemanticType, info.m_CoordinateSpace);
-        }
-        return metadata;
-    }
-
-    static inline bool IsTypeTextureType(Type type)
-    {
-        return type == TYPE_SAMPLER_2D ||
-               type == TYPE_SAMPLER_2D_ARRAY ||
-               type == TYPE_TEXTURE_2D ||
-               type == TYPE_TEXTURE_2D_ARRAY ||
-               type == TYPE_IMAGE_2D ||
-
-               type == TYPE_SAMPLER_3D ||
-               type == TYPE_SAMPLER_3D_ARRAY ||
-               type == TYPE_TEXTURE_3D ||
-               type == TYPE_IMAGE_3D ||
-
-               type == TYPE_SAMPLER_CUBE ||
-               type == TYPE_TEXTURE_CUBE;
-    }
-
-    static inline bool IsTextureType3D(TextureType type)
-    {
-        return type == TEXTURE_TYPE_3D || type == TEXTURE_TYPE_IMAGE_3D;
-    }
-
-    static inline uint32_t GetLayerCount(TextureType type)
-    {
-        return type == TEXTURE_TYPE_CUBE_MAP ? 6 : 1;
-    }
-
-    static inline TextureType TypeToTextureType(Type type)
-    {
-        switch(type)
-        {
-            case TYPE_SAMPLER_2D:       return TEXTURE_TYPE_2D;
-            case TYPE_SAMPLER_3D:       return TEXTURE_TYPE_3D;
-            case TYPE_SAMPLER_2D_ARRAY: return TEXTURE_TYPE_2D_ARRAY;
-            case TYPE_SAMPLER_CUBE:     return TEXTURE_TYPE_CUBE_MAP;
-            case TYPE_IMAGE_2D:         return TEXTURE_TYPE_IMAGE_2D;
-            case TYPE_IMAGE_3D:         return TEXTURE_TYPE_IMAGE_3D;
-            case TYPE_TEXTURE_2D:       return TEXTURE_TYPE_TEXTURE_2D;
-            case TYPE_TEXTURE_3D:       return TEXTURE_TYPE_TEXTURE_3D;
-            case TYPE_TEXTURE_2D_ARRAY: return TEXTURE_TYPE_TEXTURE_2D_ARRAY;
-            case TYPE_TEXTURE_CUBE:     return TEXTURE_TYPE_TEXTURE_CUBE;
-            case TYPE_SAMPLER:          return TEXTURE_TYPE_SAMPLER;
-            default:break;
-        }
-        assert(0);
-        return (TextureType) -1;
-    }
-
-    static inline uint32_t VectorTypeToElementCount(VertexAttribute::VectorType vector_type)
-    {
-        switch(vector_type)
-        {
-            case VertexAttribute::VECTOR_TYPE_SCALAR: return 1;
-            case VertexAttribute::VECTOR_TYPE_VEC2:   return 2;
-            case VertexAttribute::VECTOR_TYPE_VEC3:   return 3;
-            case VertexAttribute::VECTOR_TYPE_VEC4:   return 4;
-            case VertexAttribute::VECTOR_TYPE_MAT2:   return 4;
-            case VertexAttribute::VECTOR_TYPE_MAT3:   return 9;
-            case VertexAttribute::VECTOR_TYPE_MAT4:   return 16;
-        }
-        return 0;
-    }
-
-    static inline uint32_t DataTypeToByteWidth(VertexAttribute::DataType data_type)
-    {
-        switch(data_type)
-        {
-            case dmGraphics::VertexAttribute::TYPE_BYTE:           return 1;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_BYTE:  return 1;
-            case dmGraphics::VertexAttribute::TYPE_SHORT:          return 2;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_SHORT: return 2;
-            case dmGraphics::VertexAttribute::TYPE_INT:            return 4;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_INT:   return 4;
-            case dmGraphics::VertexAttribute::TYPE_FLOAT:          return 4;
-            default: break;
-        }
-        return 0;
-    }
 }
-
-
 
 #endif // #ifndef DM_GRAPHICS_UTIL_H
