+++ engine/graphics/src/null/graphics_null.cpp
@@ -27,7 +27,6 @@
 
 #include "../graphics_private.h"
 #include "../graphics_native.h"
-#include "../graphics_util.h"
 #include "../graphics_adapter.h"
 #include "graphics_null_private.h"
