+++ engine/graphics/src/vulkan/graphics_vulkan.cpp
@@ -24,7 +24,6 @@
 
 #include "../graphics_private.h"
 #include "../graphics_native.h"
-#include "../graphics_util.h"
 #include "../graphics_adapter.h"
 #include "graphics_vulkan_defines.h"
 #include "graphics_vulkan_private.h"
