+++ engine/particle/src/particle.cpp
@@ -23,7 +23,6 @@
 #include <dlib/vmath.h>
 #include <dlib/profile.h>
 #include <dlib/time.h>
-#include <graphics/graphics_util.h>
 #include <dmsdk/dlib/vmath.h>
 
 #include "particle.h"
