+++ engine/particle/src/test/test_particle.cpp
@@ -12,10 +12,6 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#if defined(_WIN32)
-#include <winsock2.h>
-#endif
-
 #define JC_TEST_IMPLEMENTATION
 #include <jc_test/jc_test.h>
 
@@ -31,7 +27,6 @@
 #include <ddf/ddf.h>
 
 #include <graphics/graphics_ddf.h>
-#include <graphics/graphics_util.h>
 
 #include "../particle.h"
 #include "../particle_private.h"
