+++ engine/render/src/render/program_utils.cpp
@@ -15,8 +15,6 @@
 #include <dlib/log.h>
 #include <dlib/dstrings.h>
 
-#include <graphics/graphics_util.h>
-
 #include "render.h"
 #include "render_private.h"
